refactor: remove the sort functionality
